Updates kthoom source code to match https://github.com/codedread/kthoom
This kthoom update fixes issues with certain cbz files that are created
with an info section: https://github.com/codedread/kthoom/issues/36

The required kthoom commit is:
https://github.com/codedread/kthoom/commit/612f3768354608bfc9048ba21ccd7cf1100b4463

Now, to update kthoom, you simply copy in the code directory from the
kthoom repo to the static/js/kthoom directory. And also copy the
kthoom.css from the code directory to the static/css/kthoom.css.

The static/css/readcbr.css file provides particular overrides for the
main styling of calibre-web to match the expected style and to override
the kthoom styling.

This uses JavaScript modules and so currently breaks compatibility with
browsers that do not support it. It will need to be compiled to a single
packed file and served with defer/nomodule to target those browsers.